requirements.txt
streamlit
google-generativeai
pypdf2
stripe
python-dotenv
